pyproject.toml
[project]
name = "aieffect"
version = "0.1.0"
description = "AI 芯片验证效率集成平台"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "junitparser>=3.0",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[tool.setuptools.packages.find]
include = ["framework*"]

[project.scripts]
aieffect = "framework.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "framework.web.app"
disallow_untyped_defs = false
warn_return_any = false
